Collect separate mount info from separate units
When migrating between NFS units (for redeployment or for vertical
scaling), there will be a period when an old unit still exists but
doesn't publish data on the relation.  Using the combined
`relation.joined_units` view makes it impossible to support this kind of
migration, as the old unit's `private-address` will always win.

The reactive documentation advises that it's normally better to iterate
over each unit and handle its data individually, and that indeed behaves
better in this case.  Nothing changes in the common case where there's
only a single NFS unit, and the requiring charm already typically has to
pick out the first response that meets its needs so isn't likely to be
confused by this change.